fix(integration): Rename package tbp -> thebrewingproject
