+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/BundledCompatibility.java
@@ -35,7 +35,7 @@ import org.betonquest.betonquest.compatibility.quests.QuestsIntegrator;
 import org.betonquest.betonquest.compatibility.redischat.RedisChatIntegrator;
 import org.betonquest.betonquest.compatibility.shopkeepers.ShopkeepersIntegrator;
 import org.betonquest.betonquest.compatibility.skript.SkriptIntegrator;
-import org.betonquest.betonquest.compatibility.tbp.TbpIntegrator;
+import org.betonquest.betonquest.compatibility.thebrewingproject.TbpIntegrator;
 import org.betonquest.betonquest.compatibility.traincarts.TrainCartsIntegrator;
 import org.betonquest.betonquest.compatibility.vault.VaultIntegrator;
 import org.betonquest.betonquest.compatibility.worldedit.WorldEditIntegrator;
