--- code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/tbp/condition/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The brewing project integration conditions.
- */
-package org.betonquest.betonquest.compatibility.tbp.condition;
