--- code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/tbp/item/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The brewing project integration brew item.
- */
-package org.betonquest.betonquest.compatibility.tbp.item;
