--- code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/tbp/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The brewing project integration.
- */
-package org.betonquest.betonquest.compatibility.tbp;
