+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/action/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The brewing project integration actions.
+ */
+package org.betonquest.betonquest.compatibility.thebrewingproject.action;
