+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/condition/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The brewing project integration conditions.
+ */
+package org.betonquest.betonquest.compatibility.thebrewingproject.condition;
