+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/item/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The brewing project integration brew item.
+ */
+package org.betonquest.betonquest.compatibility.thebrewingproject.item;
