+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The brewing project integration.
+ */
+package org.betonquest.betonquest.compatibility.thebrewingproject;
